GCP/02_Code/02_CloudRun2/requirements.txt
streamlit
google-cloud-bigquery
pandas
plotly
streamlit-option-menu
db-dtypes
